# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SponsorshipOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user and creator profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithProfile"
        "401":
          description: Unauthorized
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Email/password login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/google/callback:
    get:
      operationId: googleAuthCallback
      tags: [auth]
      summary: Google OAuth callback
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect after auth
  /auth/onboarding:
    post:
      operationId: completeOnboarding
      tags: [auth]
      summary: Save onboarding data and create creator profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingBody"
      responses:
        "200":
          description: Onboarding complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorProfile"

  # DASHBOARD
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
      required: [email, password, name]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
        needsOnboarding:
          type: boolean
      required: [token, user, needsOnboarding]
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        image:
          type: string
        createdAt:
          type: string
      required: [id, email]
    UserWithProfile:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        profile:
          $ref: "#/components/schemas/CreatorProfile"
        needsOnboarding:
          type: boolean
      required: [user, needsOnboarding]
    CreatorProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        businessName:
          type: string
        country:
          type: string
        defaultCurrency:
          type: string
        taxReservePercent:
          type: integer
        paymentTermsDays:
          type: integer
        niche:
          type: string
        averageSponsorshipValue:
          type: string
        subscriptionPlan:
          type: string
          enum: [starter, pro, agency]
        subscriptionStatus:
          type: string
        createdAt:
          type: string
      required: [id, userId, subscriptionPlan]
    OnboardingBody:
      type: object
      properties:
        name:
          type: string
        businessName:
          type: string
        country:
          type: string
        defaultCurrency:
          type: string
        paymentTermsDays:
          type: integer
        niche:
          type: string
        taxReservePercent:
          type: integer
        averageSponsorshipValue:
          type: string
      required: [name, businessName, country]
...